requirements.txt
python-dotenv
autogen-agentchat
autogen-ext[openai,azure]
openai
google-genai
toml